Match backend with frontend for valid usernames.
The frontend regex pattern is ^[a-zA-Z0-9]{2,}. We now enforce
this also in the user schema.

However there are some usernames already in the system not
matching this pattern. To deal with this we provide a
script utils/userdb.py (that has to be run once) that finds the
list of legacy usernames and stores it in the key-value store.
The new user schema continues to allow these legacy usernames.

While working on this we discovered a bug in vtjson: the schema
union() which should match nothing, matched in fact everything.
So vtjson should be upgraded to 2.2.8.

We also discovered that the error messages from KeyValueStore
objects were not very friendly. So we now do stricter input
verification.

Some new unittests are provided for the new user schema and
the added input verification for the kvstore.

+++ server/fishtest/kvstore.py
@@ -2,7 +2,7 @@ from collections.abc import MutableMapping
 from datetime import UTC
 
 from bson.codec_options import CodecOptions
-from fishtest.schemas import kvstore_schema
+from bson.errors import InvalidDocument
 from pymongo import MongoClient
 from vtjson import validate
 
@@ -19,11 +19,20 @@ class KeyValueStore(MutableMapping):
         self.__kvstore = db[collection]
 
     def __setitem__(self, key, value):
+        if not isinstance(key, str):
+            raise ValueError(f"The key {key} is not a string")
         document = {"_id": key, "value": value}
-        validate(kvstore_schema, document)
-        self.__kvstore.replace_one({"_id": key}, document, upsert=True)
+        from fishtest.schemas import kvstore_schema  # circular import issue
+
+        validate(kvstore_schema, document)  # assertion!
+        try:
+            self.__kvstore.replace_one({"_id": key}, document, upsert=True)
+        except InvalidDocument:
+            raise ValueError(f"The value {value} cannot be converted to bson") from None
 
     def __getitem__(self, key):
+        if not isinstance(key, str):
+            raise ValueError(f"The key {key} is not a string")
         document = self.__kvstore.find_one({"_id": key})
         if document is None:
             raise KeyError(key)
@@ -31,6 +40,8 @@ class KeyValueStore(MutableMapping):
             return document["value"]
 
     def __delitem__(self, key):
+        if not isinstance(key, str):
+            raise ValueError(f"The key {key} is not a string")
         d = self.__kvstore.delete_one({"_id": key})
         if d.deleted_count == 0:
             raise KeyError(key)
